move mt-bench logic from the entrypoint
